# go api: remove regex replace when decompressing