@zdylag 2024-10-11
@MaT1g3R something that is not clear to me:
Are nested compression strings allowed?
EG: `love|the|slay &1||I &0 &0 &2 relics and &2 spire` -> `I love love slay the relics and slay the spire`? If so this approach (as is) does not work, and we should add ^ as a test case.