@zdylag 2024-10-11
I made a PR to add the test cases just in case: https://github.com/Spireblight/STR-Backend/pull/21