@MaT1g3R 2025-07-18
closing because the format of the message will be fundamentally changed